Refactor List class and clean up User class
Changed the `items` array in `List` to `final` and added `this` for clarity in both `add` and `get` methods. Removed redundant code from `User` class that duplicated functionality in `List`. Added comments and documentation in `_3_A_Poor_Solution.java` to highlight the problems with raw types and the need for generics.